Agents/requirements.txt
fastapi
uvicorn
gunicorn
azure-identity
azure-ai-inference
azure-monitor-opentelemetry
opentelemetry-api
requests
